requirements.txt
requests
Pillow
ttkbootstrap
